backend/code/055/pyproject.toml
[project]
name = "mve-055-wy-allscan-sp"
version = "0.1.0"
description = "WY-factored All-Scan for Gated DeltaNet Sequence Parallelism"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4.0",
    "numpy<2",
    "pyyaml>=6.0",
    "wandb>=0.15.0",
    "modal>=1.3.0",
]
